% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleCenter.R
\docType{class}
\name{DoubleCenter-class}
\alias{DoubleCenter-class}
\title{Title}
\value{
TODO
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{row_means}}{numeric.}

\item{\code{col_means}}{numeric.}

\item{\code{overall_mean}}{numeric.}
}}

