% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleCenter.R
\name{DoubleCenter}
\alias{DoubleCenter}
\alias{transform,DoubleCenter,sparseMatrix-method}
\alias{inverse_transform,DoubleCenter,sparseLRMatrix-method}
\alias{inverse_transform,DoubleCenter,vsp_fa-method}
\title{Title}
\usage{
DoubleCenter(A)

\S4method{transform}{DoubleCenter,sparseMatrix}(iform, A)

\S4method{inverse_transform}{DoubleCenter,sparseLRMatrix}(iform, A)

\S4method{inverse_transform}{DoubleCenter,vsp_fa}(iform, A)
}
\arguments{
\item{A}{TODO}

\item{iform}{TODO}
}
\value{
\code{DoubleCenter} creates a \linkS4class{DoubleCenter}.
\itemize{
\item \code{transform()} applies the double centering operation to a matrix
}
}
\description{
Convience constructor for \linkS4class{DoubleCenter}
}
\examples{

library(igraph)
library(igraphdata)

data("karate", package = "igraphdata")

A <- get.adjacency(karate)

iform <- DoubleCenter(A)

A_tilde <- transform(iform, A)
A_tilde

A_recovered <- inverse_transform(iform, A_tilde)

all.equal(A, A_recovered)

}
