% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-generics.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Apply the inverse of an invertible transformation}
\usage{
inverse_transform(iform, A)
}
\arguments{
\item{iform}{An \linkS4class{Invertiform} object describing the transformation.}

\item{A}{A matrix to inverse transform.}
}
\value{
The inverse transformed matrix.
}
\description{
Apply the inverse of an invertible transformation
}
