% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-generics.R
\docType{class}
\name{Invertiform-class}
\alias{Invertiform-class}
\title{An abstract S4 class representing an invertible transformation}
\description{
An abstract S4 class representing an invertible transformation
}
