% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertiforms-package.R
\docType{package}
\name{invertiforms-package}
\alias{invertiforms}
\alias{invertiforms-package}
\title{invertiforms: Intervible transforms for matrix valued data}
\description{
Provides composable invertible transforms for
    (sparse) matrices.
}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

}
\keyword{internal}
