% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizedLaplacian.R
\name{NormalizedLaplacian}
\alias{NormalizedLaplacian}
\alias{transform,NormalizedLaplacian,sparseMatrix-method}
\alias{inverse_transform,NormalizedLaplacian,sparseMatrix-method}
\title{Title}
\usage{
NormalizedLaplacian(A)

\S4method{transform}{NormalizedLaplacian,sparseMatrix}(iform, A)

\S4method{inverse_transform}{NormalizedLaplacian,sparseMatrix}(iform, A)
}
\arguments{
\item{A}{TODO}

\item{iform}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

library(igraph)
library(igraphdata)

data("karate", package = "igraphdata")

A <- get.adjacency(karate)

iform <- NormalizedLaplacian(A)

L <- transform(iform, A)
L

A_recovered <- inverse_transform(iform, L)

all.equal(A, A_recovered)

}
