% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbedLaplacian.R
\docType{class}
\name{PerturbedLaplacian-class}
\alias{PerturbedLaplacian-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{tau_row}}{numeric.}

\item{\code{tau_col}}{numeric.}

\item{\code{rsA}}{numeric.}

\item{\code{csA}}{numeric.}

\item{\code{tau_choice_row}}{character.}

\item{\code{tau_choice_col}}{character.}
}}

