% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbedLaplacian.R
\name{PerturbedLaplacian}
\alias{PerturbedLaplacian}
\alias{transform,PerturbedLaplacian,sparseMatrix-method}
\alias{inverse_transform,PerturbedLaplacian,sparseLRMatrix-method}
\title{Title}
\usage{
PerturbedLaplacian(A, tau = NULL)

\S4method{transform}{PerturbedLaplacian,sparseMatrix}(iform, A)

\S4method{inverse_transform}{PerturbedLaplacian,sparseLRMatrix}(iform, A)
}
\arguments{
\item{A}{TODO}

\item{iform}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

library(igraph)
library(igraphdata)

data("karate", package = "igraphdata")

A <- get.adjacency(karate)

iform <- PerturbedLaplacian(A)

L <- transform(iform, A)
L

A_recovered <- inverse_transform(iform, L)

all.equal(A, A_recovered)

}
