% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLaplacian.R
\docType{class}
\name{RegularizedLaplacian-class}
\alias{RegularizedLaplacian-class}
\title{Regularized graph Laplacian transformation}
\description{
Regularized graph Laplacian transformation
}
\section{Slots}{

\describe{
\item{\code{tau_row}}{numeric.}

\item{\code{tau_col}}{numeric.}

\item{\code{rsA}}{numeric.}

\item{\code{csA}}{numeric.}

\item{\code{tau_choice_row}}{character.}

\item{\code{tau_choice_col}}{character.}
}}

