% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLaplacian.R
\name{RegularizedLaplacian}
\alias{RegularizedLaplacian}
\alias{transform,RegularizedLaplacian,sparseMatrix-method}
\alias{inverse_transform,RegularizedLaplacian,sparseMatrix-method}
\title{Title}
\usage{
RegularizedLaplacian(A, tau_row = NULL, tau_col = NULL)

\S4method{transform}{RegularizedLaplacian,sparseMatrix}(iform, A)

\S4method{inverse_transform}{RegularizedLaplacian,sparseMatrix}(iform, A)
}
\arguments{
\item{A}{TODO}

\item{iform}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

library(igraph)
library(igraphdata)

data("karate", package = "igraphdata")

A <- get.adjacency(karate)

iform <- RegularizedLaplacian(A)

L <- transform(iform, A)
L

A_recovered <- inverse_transform(iform, L)

all.equal(A, A_recovered)

}
