% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-generics.R
\name{transform}
\alias{transform}
\title{Apply an invertible transformation}
\usage{
transform(iform, A)
}
\arguments{
\item{iform}{An \linkS4class{Invertiform} object describing the transformation.}

\item{A}{A matrix to transform.}
}
\value{
The transformed matrix.
}
\description{
Apply an invertible transformation
}
