% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleCenter.R
\docType{class}
\name{DoubleCenter-class}
\alias{DoubleCenter-class}
\title{Row and column centering transformation}
\description{
Row and column centering transformation
}
\section{Slots}{

\describe{
\item{\code{row_means}}{numeric.}

\item{\code{col_means}}{numeric.}

\item{\code{overall_mean}}{numeric.}
}}

