% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertiforms-package.R
\docType{package}
\name{invertiforms-package}
\alias{invertiforms}
\alias{invertiforms-package}
\title{invertiforms: Invertible Transforms for Matrices}
\description{
Provides composable invertible transforms for (sparse) matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rohelab.github.io/invertiforms/}
  \item \url{https://github.com/RoheLab/invertiforms}
  \item Report bugs at \url{https://github.com/RoheLab/invertiforms/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID}) [copyright holder]

}
\keyword{internal}
