% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizedLaplacian.R
\docType{class}
\name{NormalizedLaplacian-class}
\alias{NormalizedLaplacian-class}
\title{Normalized graph Laplacian transformation}
\description{
Normalized graph Laplacian transformation
}
\section{Slots}{

\describe{
\item{\code{rsA}}{numeric.}

\item{\code{csA}}{numeric.}
}}

