% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbedLaplacian.R
\docType{class}
\name{PerturbedLaplacian-class}
\alias{PerturbedLaplacian-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{tau}}{numeric.}

\item{\code{rsA}}{numeric.}

\item{\code{csA}}{numeric.}

\item{\code{tau_choice}}{character.}
}}

