% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegularizedLaplacian.R
\name{RegularizedLaplacian}
\alias{RegularizedLaplacian}
\alias{transform,RegularizedLaplacian,Matrix-method}
\alias{transform,RegularizedLaplacian,matrix-method}
\alias{transform,RegularizedLaplacian,sparseLRMatrix-method}
\alias{inverse_transform,RegularizedLaplacian,Matrix-method}
\alias{inverse_transform,RegularizedLaplacian,matrix-method}
\alias{inverse_transform,RegularizedLaplacian,vsp_fa-method}
\title{Title}
\usage{
RegularizedLaplacian(A, tau_row = NULL, tau_col = NULL)

\S4method{transform}{RegularizedLaplacian,Matrix}(iform, A)

\S4method{transform}{RegularizedLaplacian,matrix}(iform, A)

\S4method{transform}{RegularizedLaplacian,sparseLRMatrix}(iform, A)

\S4method{inverse_transform}{RegularizedLaplacian,Matrix}(iform, A)

\S4method{inverse_transform}{RegularizedLaplacian,matrix}(iform, A)

\S4method{inverse_transform}{RegularizedLaplacian,vsp_fa}(iform, A)
}
\arguments{
\item{A}{A \code{\link[Matrix:Matrix]{Matrix::Matrix()}} or \code{\link[sparseLRMatrix:sparseLRMatrix]{sparseLRMatrix::sparseLRMatrix()}} object.}

\item{tau_row}{TODO}

\item{tau_col}{TODO}

\item{iform}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{

library(igraph)
library(igraphdata)

data("karate", package = "igraphdata")

A <- get.adjacency(karate)

iform <- RegularizedLaplacian(A)

L <- transform(iform, A)
L

A_recovered <- inverse_transform(iform, L)

all.equal(A, A_recovered)

}
